+++ src/test/java/io/github/classgraph/issues/issue238/Issue238Test.java
@@ -89,7 +89,7 @@ public class Issue238Test {
                 .scan()) {
             final var classNames = scanResult.getAllClasses().get(E.class.getName()).getAllSuperclasses().getNames();
             assertThat(classNames).containsExactly(F.class.getName(), A.class.getName(), G.class.getName(),
-                    B.class.getName(), D.class.getName(), C.class.getName());
+                    B.class.getName(), D.class.getName(), C.class.getName(), "java.lang.Object");
         }
     }
 }